Revert "Merge branch 'master' of https://github.com/jramosv/sepweb"
This reverts commit 9a1bb6afd1afacd48a6301cb045f805c833cf7dc, reversing
changes made to 3ec4553d7a4e3922c727199dbba2ef7e809e5456.

+++ app/Http/Controllers/PatientController.php
@@ -129,12 +129,15 @@ class PatientController extends Controller
      * @param  \App\Patient  $patient
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Patient $patient)
+    public function destroy(Request $request, $id)
 {
 
-    $patient->delete();
-    return redirect('/pacientes')->with('status', 'La enfermera se elimino de forma permanente!');
-  
+    if($request->ajax()):
+            $patient = \App\Patient::find($request->$id);
+            $patient->delete();
+    endif;
+
+    return redirect()->to('/pacientes');
 
 }
 }
